Centered the text in the textbox. Added some room from the text to the edge of the box
Removed background color

Added a border to one of the pages too see how it looks. Refactored the code

Changed the titles on the 'Om Oss' page. Changed so that the nav-bar is on the left-side of the page, and the information on the right-hand side.

Changed so that the text automatically changes color depending on what 'light mode' the user is in

Removed the uneccessary files from the public folder

removed unneccesary files